% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-model.R
\name{add_variable}
\alias{add_variable}
\title{Adds a variable to the model}
\usage{
add_variable(model, variable, type = "continuous", lb = -Inf, ub = Inf,
  ...)
}
\arguments{
\item{model}{the model}

\item{variable}{the variable name/definition}

\item{type}{must be either continuous, integer or binary}

\item{lb}{the lower bound of the variable}

\item{ub}{the upper bound of the variable}

\item{...}{quantifiers for the indexed variabled}
}
\description{
A variable can either by a name or an indexed name. See examples.
}
\examples{
library(magrittr)
MIPModel() \%>\%
 add_variable(x) \%>\% # creates 1 variable named x
 add_variable(y[i], i = 1:10) # creates 10 variables

}

