% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milp-linearopt-variables.R
\name{as_colwise}
\alias{as_colwise}
\title{As_colwise}
\usage{
as_colwise(x)
}
\arguments{
\item{x}{a list of numeric vectors or a numeric vector}
}
\description{
Convert lists or vectors to colwise semantic.
}
