% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{CBCsolution}
\alias{CBCsolution}
\title{Sends a model to solve to CBC and gets the solution back}
\usage{
CBCsolution(model, modelFile = "Model.mps", saveSolution = 1,
  solutionFile = "cbcsolution.txt", modelName = "ompr1", timeLimit, ...)
}
\arguments{
\item{model}{The ompr model to solve}

\item{modelFile}{The file where the model should be stored (Model.mps by default)}

\item{saveSolution}{Boolean to save the solution in a specific file}

\item{solutionFile}{The file containing the solution (cbcsolution.txt by default)}

\item{modelName}{If the model should have an specific name inside the mps file (ompr1 by default)}

\item{timeLimit}{Time limit for the solver}

\item{...}{More cbc argumets (to be implemented)}
}
\description{
It uses the write_MPS, solve_CBC and read_CBCSolution functions in sequence with some default values
}
\examples{
library(magrittr)
result <- MIPModel() \%>\%
 add_variable(x, type = "integer") \%>\%
 add_variable(y, type = "continuous", lb = 0) \%>\%
 set_bounds(x, lb = 0) \%>\%
 set_objective(x + y, "max") \%>\%
 add_constraint(x + y <= 11.25)

solution <- CBCsolution(model=result, timeLimit=60)

}

