% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{colwise}
\alias{colwise}
\title{Cowise}
\usage{
colwise(...)
}
\arguments{
\item{...}{create a colwise vector}
}
\description{
Cowise
}
\examples{
\dontrun{
  # vectors create matrix rows
  # x[1, 1]
  # x[2, 1]
  # x[3, 1]
  x[1:3, 1]

  # colwise() creates columns per row
  # 1 * x[1, 1] + 2 * x[1, 2] + 3 * x[1, 3]
  colwise(1, 2, 3) * x[1, colwise(1, 2, 3)]

  # you can also combine the two
  # x[1, 1]
  # x[2, 1] + x[2, 2]
  # x[3, 1] + x[3, 2] + x[3, 2]
  x[1:3, colwise(1, 1:2, 1:3)]
}
}
