% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-glpk.R
\name{configure_glpk_solver}
\alias{configure_glpk_solver}
\title{This function constructs a glpk solver}
\usage{
configure_glpk_solver(verbose = FALSE)
}
\arguments{
\item{verbose}{would you like to see GLPK messages?}
}
\description{
This function constructs a glpk solver
}

