% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-optimization-model-impl.R
\name{extract_bound_from_constraint}
\alias{extract_bound_from_constraint}
\alias{extract_bound_from_constraint,LinearFunction,numeric,LinearConstraintSense-method}
\alias{extract_bound_from_constraint,LinearTerm,numeric,LinearConstraintSense-method}
\title{Extract a scalar bound for a variable from a constraint}
\usage{
extract_bound_from_constraint(lhs, rhs, sense)

\S4method{extract_bound_from_constraint}{LinearFunction,numeric,LinearConstraintSense}(lhs, rhs, sense)

\S4method{extract_bound_from_constraint}{LinearTerm,numeric,LinearConstraintSense}(lhs, rhs, sense)
}
\description{
Extract a scalar bound for a variable from a constraint
}
\keyword{internal}
