% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_coefficients}
\alias{extract_coefficients}
\title{Extracts the coefficients out of an ast}
\usage{
extract_coefficients(ast)
}
\arguments{
\item{ast}{the ast}
}
\value{
a list with values for constants and coefficients
}
\description{
Should only be used if you intentd to develop your own solver.
}
\details{
The expression should be a simple sum.
Only multiplication and + operators are allowed.
For each multiplication operation, one operand must be a numeric
the other a non-numeric.
}
\examples{
extract_coefficients(substitute(5 * x + 1))

}

