% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-solution.R
\name{get_column_duals}
\alias{get_column_duals}
\title{Gets the column duals of a solution}
\usage{
get_column_duals(solution)
}
\arguments{
\item{solution}{a solution}
}
\value{
Either a numeric vector with one element per column or `NA_real_`.
}
\description{
Gets the column duals of a solution
}
