% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-solution.R
\name{get_row_duals}
\alias{get_row_duals}
\title{Gets the row duals of a solution}
\usage{
get_row_duals(solution)
}
\arguments{
\item{solution}{a solution}
}
\value{
Either a numeric vector with one element per row or `NA_real_`.
}
\description{
Gets the row duals of a solution
}
