% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-solution.R
\name{get_solution}
\alias{get_solution}
\title{Extracts variable values from a solution}
\usage{
get_solution(solution, exp)
}
\arguments{
\item{solution}{the solution object}

\item{exp}{a variable expression. You can partially bind indexes.}
}
\value{
a data.frame. One row for each variable instance
        and a column for each index.
}
\description{
Extracts variable values from a solution
}

