% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-model.R
\name{is_defined}
\alias{is_defined}
\title{Check if variable is defined in model}
\usage{
is_defined(model, variable)
}
\arguments{
\item{model}{a model object}

\item{variable}{the variable expression to check}
}
\value{
TRUE iff the variable is part of the model
}
\description{
Check if variable is defined in model
}

