% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-math.R
\name{latex_math}
\alias{latex_math}
\alias{knitr_math}
\alias{latex_constraints}
\alias{knitr_constraints}
\alias{latex_objective}
\alias{knitr_objective}
\title{Output expression as LaTeX math}
\usage{
latex_math(x, fmt = "\%s")

knitr_math(x, fmt = "\\\\[\%s\\\\]", ...)

latex_constraints(model, fmt = "\%s", collapse = ", ")

knitr_constraints(model, fmt = "\\\\[\%s\\\\]", collapse = ", ", ...)

latex_objective(model, expand = FALSE, sense = TRUE, fmt = "\%s")

knitr_objective(model, expand = FALSE, sense = TRUE,
  fmt = "\\\\[\%s\\\\]", ...)
}
\arguments{
\item{x}{Expression or character string}

\item{fmt}{Format string, a la \code{\link[base]{sprintf}}}

\item{\dots}{Additional arguments to be passed to
\code{knitr::asis_output()}}

\item{model}{MIP model of class \code{optimization_model}}

\item{collapse}{Character string to separate constraints}

\item{expand}{Whether to process the expanded or original expression}

\item{sense}{Whether to include the model sense in the output}
}
\value{
LaTeX command as a character string for \code{latex_math()}.
  knit_asis object for \code{knitr_math()}.
}
\description{
Convert a single constraint, objective function, or some other R
expression object to a LaTeX math command. \code{knitr_math()} is a
convenience wrapper for \code{latex_math()} plus the knitr chunk
option \code{results = 'asis'}.
}
