% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-optimization-model-linear-constraints.R
\docType{class}
\name{LinearConstraintSense-class}
\alias{LinearConstraintSense-class}
\alias{LinearConstraint-class}
\alias{new_linear_constraint}
\alias{new_linear_constraint,LinearFunction,numeric,LinearConstraintSense-method}
\alias{new_linear_constraint,LinearFunction,LinearTerm,LinearConstraintSense-method}
\alias{new_linear_constraint,LinearFunction,LinearFunction,LinearConstraintSense-method}
\alias{new_linear_constraint,LinearTerm,ANY,LinearConstraintSense-method}
\alias{<=,AbstractLinearFunction,ANY-method}
\alias{==,AbstractLinearFunction,ANY-method}
\alias{>=,AbstractLinearFunction,ANY-method}
\title{Internal linear constraints classes and methods}
\usage{
new_linear_constraint(lhs, rhs, sense)

\S4method{new_linear_constraint}{LinearFunction,numeric,LinearConstraintSense}(lhs, rhs, sense)

\S4method{new_linear_constraint}{LinearFunction,LinearTerm,LinearConstraintSense}(lhs, rhs, sense)

\S4method{new_linear_constraint}{LinearFunction,LinearFunction,LinearConstraintSense}(lhs, rhs, sense)

\S4method{new_linear_constraint}{LinearTerm,ANY,LinearConstraintSense}(lhs, rhs, sense)

\S4method{<=}{AbstractLinearFunction,ANY}(e1, e2)

\S4method{==}{AbstractLinearFunction,ANY}(e1, e2)

\S4method{>=}{AbstractLinearFunction,ANY}(e1, e2)
}
\arguments{
\item{lhs}{a linear function}

\item{rhs}{a numeric scalar}

\item{sense}{the constraint sense}

\item{e1}{a parameter}

\item{e2}{a parameter}
}
\description{
Internal linear constraints classes and methods
}
\section{Slots}{

\describe{
\item{\code{sense}}{leq, geq or eq}
}}

