% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-optimization-model-linear-functions.R,
%   R/linear-optimization-model-impl.R
\docType{class}
\name{OmprLinearVariable-class}
\alias{OmprLinearVariable-class}
\alias{AbstractLinearFunction-class}
\alias{LinearTerm-class}
\alias{LinearFunction-class}
\alias{*,LinearTerm,numeric-method}
\alias{*,numeric,LinearTerm-method}
\alias{/,LinearTerm,numeric-method}
\alias{+,LinearTerm,numeric-method}
\alias{+,numeric,LinearTerm-method}
\alias{+,LinearTerm,missing-method}
\alias{+,LinearTerm,LinearTerm-method}
\alias{-,LinearTerm,missing-method}
\alias{+,LinearFunction,numeric-method}
\alias{+,numeric,LinearFunction-method}
\alias{+,LinearFunction,LinearTerm-method}
\alias{+,LinearTerm,LinearFunction-method}
\alias{+,LinearFunction,LinearFunction-method}
\alias{-,LinearFunction,numeric-method}
\alias{-,numeric,LinearFunction-method}
\alias{-,LinearTerm,LinearTerm-method}
\alias{-,numeric,LinearTerm-method}
\alias{-,LinearFunction,LinearTerm-method}
\alias{-,LinearTerm,LinearFunction-method}
\alias{-,LinearFunction,LinearFunction-method}
\alias{*,LinearFunction,numeric-method}
\alias{*,numeric,LinearFunction-method}
\alias{/,LinearFunction,numeric-method}
\alias{OmprLinearVariableCollection-class}
\alias{[,OmprLinearVariableCollection,ANY,ANY,missing-method}
\alias{length,OmprLinearVariableCollection-method}
\alias{length,OmprLinearVariable-method}
\title{Internal linear function classes and methods}
\usage{
\S4method{*}{LinearTerm,numeric}(e1, e2)

\S4method{*}{numeric,LinearTerm}(e1, e2)

\S4method{/}{LinearTerm,numeric}(e1, e2)

\S4method{+}{LinearTerm,numeric}(e1, e2)

\S4method{+}{numeric,LinearTerm}(e1, e2)

\S4method{+}{LinearTerm,missing}(e1, e2)

\S4method{+}{LinearTerm,LinearTerm}(e1, e2)

\S4method{-}{LinearTerm,missing}(e1, e2)

\S4method{+}{LinearFunction,numeric}(e1, e2)

\S4method{+}{numeric,LinearFunction}(e1, e2)

\S4method{+}{LinearFunction,LinearTerm}(e1, e2)

\S4method{+}{LinearTerm,LinearFunction}(e1, e2)

\S4method{+}{LinearFunction,LinearFunction}(e1, e2)

\S4method{-}{LinearFunction,numeric}(e1, e2)

\S4method{-}{numeric,LinearFunction}(e1, e2)

\S4method{-}{LinearTerm,LinearTerm}(e1, e2)

\S4method{-}{numeric,LinearTerm}(e1, e2)

\S4method{-}{LinearFunction,LinearTerm}(e1, e2)

\S4method{-}{LinearTerm,LinearFunction}(e1, e2)

\S4method{-}{LinearFunction,LinearFunction}(e1, e2)

\S4method{-}{LinearFunction,LinearFunction}(e1, e2)

\S4method{*}{LinearFunction,numeric}(e1, e2)

\S4method{*}{numeric,LinearFunction}(e1, e2)

\S4method{/}{LinearFunction,numeric}(e1, e2)

\S4method{[}{OmprLinearVariableCollection,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{length}{OmprLinearVariableCollection}(x)

\S4method{length}{OmprLinearVariable}(x)
}
\arguments{
\item{e1}{a parameter}

\item{e2}{a parameter}

\item{x}{some value}

\item{i}{an index}

\item{j}{an index}

\item{...}{more indexes}

\item{drop}{unused}
}
\description{
Internal linear function classes and methods
}
\keyword{internal}
