% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\docType{class}
\name{LinearVariableCollection-class}
\alias{LinearVariableCollection-class}
\title{An S4 class that represents a collection of variables}
\description{
An S4 class that represents a collection of variables
}
\section{Slots}{

\describe{
\item{\code{variables}}{a data frame hold the variable coefficients. One line for reach variable, row and column.}

\item{\code{index_mapping}}{a function that takes a variable name as character and returns a mapping table that maps column ids to variable indexes.}
}}

