% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{-,LinearVariableCollection,LinearVariableCollection-method}
\alias{-,LinearVariableCollection,LinearVariableCollection-method}
\title{Minus}
\usage{
\S4method{-}{LinearVariableCollection,LinearVariableCollection}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableCollection'}

\item{e2}{an object of type 'LinearVariableCollection'}
}
\description{
Equivalent to `e1 + -1 * e2`
}
