% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{-,LinearVariableCollection,missing-method}
\alias{-,LinearVariableCollection,missing-method}
\title{Unary Minus}
\usage{
\S4method{-}{LinearVariableCollection,missing}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableCollection'}

\item{e2}{a missing value}
}
\description{
Equivalent to `e1 * (-1)`
}
