% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{-,LinearVariableCollection,numeric-method}
\alias{-,LinearVariableCollection,numeric-method}
\title{Plus}
\usage{
\S4method{-}{LinearVariableCollection,numeric}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableCollection'}

\item{e2}{a numeric value}
}
\description{
Equivalent to `e1 + -1 * e2`
}
