% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\docType{class}
\name{LinearVariableSum-class}
\alias{LinearVariableSum-class}
\title{Holds a sum of a constant and a linear variable collection}
\description{
Holds a sum of a constant and a linear variable collection
}
\section{Slots}{

\describe{
\item{\code{constant}}{a numeric vector}

\item{\code{variables}}{a variable collection}
}}

