% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{-,LinearVariableSum,numeric-method}
\alias{-,LinearVariableSum,numeric-method}
\title{Minus}
\usage{
\S4method{-}{LinearVariableSum,numeric}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableSum'}

\item{e2}{a numeric vector}
}
\description{
Equivalent to `e1 + -1 * e2`
}
