% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milp.R
\name{MILPModel}
\alias{MILPModel}
\title{Create a new MILP Model}
\usage{
MILPModel()
}
\description{
It will eventually replace the old `MIPModel` backend for linear models.
Use with caution and please report any bugs you find.
}
