% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-model.R
\docType{class}
\name{Model-class}
\alias{Model}
\alias{Model-class}
\title{An S4 class to represent a Model.}
\description{
An S4 class to represent a Model.
}
\section{Slots}{

\describe{
\item{\code{variables}}{a list of S4 Variable objects}

\item{\code{objective}}{the objective function}

\item{\code{constraints}}{a list of constraints}
}}

