% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{nconstraints}
\alias{nconstraints}
\title{Number of variables (rows) of the model}
\usage{
nconstraints(model)
}
\arguments{
\item{model}{the model}
}
\value{
An integer equal to the number of variables. A
variable is here a column in the resulting constraint matrix.
}
\description{
Number of variables (rows) of the model
}
\examples{
library(magrittr)
model <- MIPModel() \%>\%
  add_variable(x) \%>\%
  add_variable(y[i], i = 1:10)
nconstraints(model) # 11
}
