% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-solution.R
\name{new_solution}
\alias{new_solution}
\title{Create a new solution}
\usage{
new_solution(model, objective_value, status, solution)
}
\arguments{
\item{model}{the optimization model that was solved}

\item{objective_value}{a numeric objective value}

\item{status}{the status of the solution}

\item{solution}{a named numeric vector containing the solution values}
}
\description{
This function/class should only be used if you develop your own solver.
}
