% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{-,numeric,LinearVariableSum-method}
\alias{-,numeric,LinearVariableSum-method}
\title{Minus}
\usage{
\S4method{-}{numeric,LinearVariableSum}(e1, e2)
}
\arguments{
\item{e1}{a numeric vector}

\item{e2}{an object of type 'LinearVariableSum'}
}
\description{
Equivalent to `(-1 * e2) - (-1 * e1)`
}
