% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{objective_function}
\alias{objective_function}
\title{Extract the objective function of a model}
\usage{
objective_function(model)
}
\arguments{
\item{model}{the model}
}
\description{
Extract the objective function of a model
}

