% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-model.R
\docType{class}
\name{ObjectiveFunction-class}
\alias{ObjectiveFunction}
\alias{ObjectiveFunction-class}
\title{An S4 class to represent a ObjectiveFunction}
\description{
An S4 class to represent a ObjectiveFunction
}

