% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{+,LinearVariableCollection,LinearVariableSum-method}
\alias{+,LinearVariableCollection,LinearVariableSum-method}
\title{Plus}
\usage{
\S4method{+}{LinearVariableCollection,LinearVariableSum}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableCollection'}

\item{e2}{an object of type 'LinearVariableSum'}
}
\description{
Equivalent to `e2 + e1`
}
