% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{+,LinearVariableCollection,numeric-method}
\alias{+,LinearVariableCollection,numeric-method}
\title{Plus}
\usage{
\S4method{+}{LinearVariableCollection,numeric}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableCollection'}

\item{e2}{a numeric vector without NAs}
}
\description{
Adds a constant numeric vector to a variable. The constant needs to be a vector of length 1.
}
