% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{+,LinearVariableSum,LinearVariableSum-method}
\alias{+,LinearVariableSum,LinearVariableSum-method}
\title{Plus}
\usage{
\S4method{+}{LinearVariableSum,LinearVariableSum}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableSum'}

\item{e2}{an object of type 'LinearVariableSum'}
}
\value{
Returns an object of type 'LinearVariableSum'
}
\description{
Add two object of 'LinearVariableSum'. I.e. variables + constants
}
