% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{+,LinearVariableSum,missing-method}
\alias{+,LinearVariableSum,missing-method}
\title{Unary Plus}
\usage{
\S4method{+}{LinearVariableSum,missing}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableSum'}

\item{e2}{a missing value}
}
\description{
Equivalent to `e1`
}
