% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{read_CBCsolution}
\alias{read_CBCsolution}
\title{Reads the solution file generated by CBC}
\usage{
read_CBCsolution(solutionFile, mapping = NULL)
}
\arguments{
\item{solutionFile}{The file containing the solution.}

\item{mapping}{The mapping generated by the write_MPS function. If none is given the random
names generated during write_MPS are kept.}
}
\description{
Reads the solution file generated by CBC
}
\examples{
library(magrittr)
result <- MIPModel() \%>\%
 add_variable(x, type = "integer") \%>\%
 add_variable(y, type = "continuous", lb = 0) \%>\%
 set_bounds(x, lb = 0) \%>\%
 set_objective(x + y, "max") \%>\%
 add_constraint(x + y <= 11.25)

write_MPS(model = model, file = "test.mps", modelname = "TEST")

solve_CBC(mpsFile="test.mps", saveSolution=1, solutionFile="testsolution.txt", timeLimit=60)

read_CBCsolution("testsolution.txt")
}

