% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-model.R
\name{set_bounds}
\alias{set_bounds}
\alias{set_bounds_}
\title{Sets the bounds of a variable}
\usage{
set_bounds(model, variable, lb = NULL, ub = NULL, ...)
set_bounds_(model, variable, lb = NULL, ub = NULL, ...)
}
\arguments{
\item{model}{the model}

\item{variable}{the variable name/definition}

\item{lb}{the lower bound of the variable}

\item{ub}{the upper bound of the variable}

\item{...}{quantifiers for the indexed variabled}
}
\description{
Change the lower and upper bounds of a named variable,
indexed variable or a group of variables.
}

