% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-model.R
\docType{methods}
\name{show,Model-method}
\alias{show,Model-method}
\title{Outputs the model to the console}
\usage{
\S4method{show}{Model}(object)
}
\arguments{
\item{object}{the model object}
}
\description{
Outputs the model to the console
}

