% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-solution.R
\docType{methods}
\name{show,Solution-method}
\alias{show,Solution-method}
\title{Outputs a model summary to the console.}
\usage{
\S4method{show}{Solution}(object)
}
\arguments{
\item{object}{the solution object}
}
\description{
Outputs a model summary to the console.
}

