% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\docType{methods}
\name{/,numeric,LinearVariableSum-method}
\alias{/,numeric,LinearVariableSum-method}
\title{Division}
\usage{
\S4method{/}{numeric,LinearVariableSum}(e1, e2)
}
\arguments{
\item{e1}{a numeric value}

\item{e2}{an object of type 'LinearVariableSum'}
}
\description{
Equivalent to `e2 / e1`
}
