% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{solve_CBC}
\alias{solve_CBC}
\title{Calls cbc through a command to solve a MIP problem in a  .mps file}
\usage{
solve_CBC(mpsFile, saveSolution = 0, solutionFile = "", timeLimit = 0,
  ...)
}
\arguments{
\item{mpsFile}{The .mps file containing the model}

\item{saveSolution}{If the solution should be stored in a specific file. 0 would save it in
generic file that would be overwritten in the next solve call, 1 would save it in other file
with name given or 'randomly' generated}

\item{solutionFile}{(optional) The name to give to the solution file.}

\item{timeLimit}{(optional) The time limit to give the solver, in seconds.}

\item{...}{(optional) The rest of the parameters for the solver (to be implemented)}
}
\description{
Calls cbc through a command to solve a MIP problem in a  .mps file
}
\examples{
library(magrittr)
result <- MIPModel() \%>\%
 add_variable(x, type = "integer") \%>\%
 add_variable(y, type = "continuous", lb = 0) \%>\%
 set_bounds(x, lb = 0) \%>\%
 set_objective(x + y, "max") \%>\%
 add_constraint(x + y <= 11.25)

write_MPS(model = model, file = "test.mps", modelname = "TEST")

solve_CBC(mpsFile="test.mps", saveSolution=1, solutionFile="testsolution.txt", timeLimit=60)
}

