% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{solve_model}
\alias{solve_model}
\title{Solve a model}
\usage{
solve_model(model, solver)
}
\arguments{
\item{model}{the model}

\item{solver}{a function mapping a model to a solution}
}
\value{
solver(model)
}
\description{
Solve a model
}
