% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\docType{methods}
\name{[,LinearVariableCollection,ANY,ANY,missing-method}
\alias{[,LinearVariableCollection,ANY,ANY,missing-method}
\title{Subset}
\usage{
\S4method{[}{LinearVariableCollection,ANY,ANY,missing}(x, i, j, ...,
  drop = TRUE)
}
\arguments{
\item{x}{an object of type 'LinearVariableCollection'}

\item{i}{a numeric vector or a list of numeric vectors}

\item{j}{a numeric vector or a list of numeric vectors}

\item{...}{more a numeric vectors or a lists of numeric vectors}

\item{drop}{do not use this parameter}
}
\value{
a new object of type 'LinearVariableCollection'
}
\description{
TODO: to be documented when ready
}
