% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sum_exp}
\alias{sum_exp}
\title{Constructs a sum expression}
\usage{
sum_exp(exp, ...)
}
\arguments{
\item{exp}{an expression that can be repeated to a sum}

\item{...}{bind variables in exp using dots. See examples.}
}
\value{
the expanded sum as an AST
}
\description{
Constructs a sum expression
}
\examples{
sum_exp(x[i], i = 1:10)

}

