% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sum_expr}
\alias{sum_expr}
\title{Construct a sum expression}
\usage{
sum_expr(.expr, ...)
}
\arguments{
\item{...}{bind variables in expr using dots. See examples.}

\item{expr}{an expression that can be expanded to a sum}
}
\value{
the expanded sum as an AST
}
\description{
This functions helps to create dynamic sum expression
based on external variables. Should only be used within
other 'ompr' functions.
}
\seealso{
\code{\link{add_constraint}}

\code{\link{set_objective}}
}
