% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-optimization-model-impl.R
\name{sum_over}
\alias{sum_over}
\alias{sum_expr}
\title{Sum over indexes}
\usage{
sum_over(.expr, ...)

sum_expr(.expr, ...)
}
\arguments{
\item{.expr}{an expression that can be expanded to a sum}

\item{...}{bind variables in expr using dots. See examples.}
}
\value{
the sum over all the indexes
}
\description{
This functions helps to create summations over indexes.
}
\examples{
if (FALSE) {
# create a sum from x_1 to x_10
sum_over(x[i], i = 1:10)
# create a sum from x_2 to x_10 with even indexes
sum_over(x[i], i = 1:10, i \%\% 2 == 0)
}
}
\seealso{
\code{\link{add_constraint}}

\code{\link{set_objective}}
}
