% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{*,LinearVariableCollection,numeric-method}
\alias{*,LinearVariableCollection,numeric-method}
\title{Multiply}
\usage{
\S4method{*}{LinearVariableCollection,numeric}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableCollection'}

\item{e2}{a numeric vector}
}
\description{
Multiplies the coefficients rowwise with a given numeric vector.
If the numeric vector is a `linear_transposed_vector`, it will multiply the vector
with each variable per row.
}
