% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{*,LinearVariableSum,numeric-method}
\alias{*,LinearVariableSum,numeric-method}
\title{Multiply}
\usage{
\S4method{*}{LinearVariableSum,numeric}(e1, e2)
}
\arguments{
\item{e1}{an object of type 'LinearVariableSum'}

\item{e2}{a numeric vector}
}
\description{
It will multiply the numeric vector with both the constant and the variable in
'LinearVariableSum'
}
