% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{*,numeric,LinearVariableCollection-method}
\alias{*,numeric,LinearVariableCollection-method}
\title{Multiply}
\usage{
\S4method{*}{numeric,LinearVariableCollection}(e1, e2)
}
\arguments{
\item{e1}{a numeric value}

\item{e2}{an object of type 'LinearVariableCollection'}
}
\description{
Equivalent to `e2 * e1`
}
