% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\name{*,numeric,LinearVariableSum-method}
\alias{*,numeric,LinearVariableSum-method}
\title{Multiply}
\usage{
\S4method{*}{numeric,LinearVariableSum}(e1, e2)
}
\arguments{
\item{e1}{a numeric vector}

\item{e2}{an object of type 'LinearVariableSum'}
}
\description{
It will multiply the numeric vector with both the constant and the variable in
'LinearVariableSum'
}
