% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-variable.R
\docType{class}
\name{Variable-class}
\alias{Variable}
\alias{Variable-class}
\title{An S4 class to represent a Variable.}
\description{
An S4 class to represent a Variable.
}
\section{Slots}{

\describe{
\item{\code{arity}}{the number of subscripts the variable has}

\item{\code{type}}{the type of the variable must be one of (binary, continuous, integer). Default is continuous.}

\item{\code{lb}}{an optional lower bound. No value means unbounded.}

\item{\code{ub}}{an optional upper bound. No value means unbounded.}
}}

