% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{variable_bounds}
\alias{variable_bounds}
\title{Returns the variable lower and upper bounds}
\usage{
variable_bounds(model)
}
\arguments{
\item{model}{the model}
}
\value{
a list with two components 'lower' and 'upper' each
having a numeric vector of bounds. One for each variable.
}
\description{
Returns the variable lower and upper bounds
}

