% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{variable_types}
\alias{variable_types}
\title{Return the variable types}
\usage{
variable_types(model)
}
\arguments{
\item{model}{the model}
}
\value{
a factor with levels binary, continuous, integer
}
\description{
One component for each variable in the correct order
}

