% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-model.R
\name{add_constraint}
\alias{add_constraint}
\title{Add a constraint}
\usage{
add_constraint(model, lhs, direction, rhs, ...)
}
\arguments{
\item{model}{the model}

\item{lhs}{the linear objective as a sum of variables and constants}

\item{direction}{either "<=", ">=" or "=="}

\item{rhs}{the linear objective as a sum of variables and constants}

\item{...}{quantifiers for the indexed variables. For all combinations of
bound variables a new constraint is created.}
}
\value{
a Model with new constraints added
}
\description{
Add a constraint
}
\examples{
library(magrittr)
MIPModel() \%>\%
 add_variable(x[i], i = 1:5) \%>\%
 add_constraint(x[i], ">=", 1, i = 1:5) # creates 5 constraints

}

