% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearopt-variables.R
\docType{class}
\name{LinearVariable-class}
\alias{LinearVariable-class}
\title{An S4 class that represents a single variable}
\description{
An S4 class that represents a single variable
}
\section{Slots}{

\describe{
\item{\code{variable}}{a linear variable collection with just one index '1'}
}}

