% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{objective_function}
\alias{objective_function}
\title{Extract the objective function from a model}
\usage{
objective_function(model)
}
\arguments{
\item{model}{the model}
}
\description{
Extract the objective function from a model
}
\examples{
library(magrittr)
model <- MIPModel() \%>\%
  add_variable(x[i], i = 1:5) \%>\%
  set_objective(sum_expr(i * x[i], i = 1:5) + 10)
objective_function(model)
}
