% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution-api.R
\name{objective_value}
\alias{objective_value}
\title{Extract the numerical objective value from a solution}
\usage{
objective_value(solution)
}
\arguments{
\item{solution}{a solution}
}
\value{
numeric single item vector
}
\description{
Extract the numerical objective value from a solution
}
