% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-objective-function.R
\docType{class}
\name{ObjectiveFunction-class}
\alias{ObjectiveFunction}
\alias{ObjectiveFunction-class}
\title{An S4 class to represent a ObjectiveFunction}
\description{
An S4 class to represent a ObjectiveFunction
}
\section{Slots}{

\describe{
\item{\code{expression}}{the expression in standard form}

\item{\code{original_expression}}{the original expression as supplied by the user}

\item{\code{direction}}{the direction of optimization}
}}

