% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-solution.R
\name{solver_status}
\alias{solver_status}
\title{Get the solver status from a solution}
\usage{
solver_status(solution)
}
\arguments{
\item{solution}{a solution}
}
\value{
character vector being either "infeasible", "optimal", "unbounded", "userlimit" or "error
}
\description{
Get the solver status from a solution
}
