% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sum_expr}
\alias{sum_expr}
\title{Constructs a sum expression}
\usage{
sum_expr(exp, ...)
}
\arguments{
\item{exp}{an expression that can be expanded to a sum}

\item{...}{bind variables in exp using dots. See examples.}
}
\value{
the expanded sum as an AST
}
\description{
Constructs a sum expression
}
\examples{
sum_expr(x[i], i = 1:10)

}

