% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{variable_types}
\alias{variable_types}
\title{Variable types of a model}
\usage{
variable_types(model)
}
\arguments{
\item{model}{the model}
}
\value{
a factor with levels binary, continuous, integer
}
\description{
One component for each variable in the correct order
}
\examples{
library(magrittr)
model <- MIPModel() \%>\%
  add_variable(x, type = "binary") \%>\%
  add_variable(y, type = "continuous") \%>\%
  add_variable(z, type = "integer")
variable_types(model)
}
