% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{write_MPS}
\alias{write_MPS}
\title{Write the model as a .mps file}
\usage{
write_MPS(model, file, modelname = "OMPR1")
}
\arguments{
\item{model}{the model}

\item{file}{the route and name of the output file}

\item{modelname}{(optional). Sets the model name inside the file, default is "OMPR1"}
}
\value{
It writes the model to the indicated file
}
\description{
Write the model as a .mps file
}
\examples{
library(magrittr)
result <- MIPModel() \%>\%
 add_variable(x, type = "integer") \%>\%
 add_variable(y, type = "continuous", lb = 0) \%>\%
 set_bounds(x, lb = 0) \%>\%
 set_objective(x + y, "max") \%>\%
 add_constraint(x + y <= 11.25)

write_MPS(model = model, file = "test.mps", modelname = "TEST")
}

